split apart model selection routines from paper-specific script
