+++ experiments/model_configs.py
@@ -26,9 +26,10 @@ ModelConfig = namedtuple(
         "n_epochs",
         "dropout_probability",
         "max_ic50",
+        "minibatch_size",
     ])
 
-HIDDEN1_LAYER_SIZES = [
+HIDDEN_LAYER_SIZES = [
     50,
     400,
 ]
@@ -41,6 +42,7 @@ INITILIZATION_METHODS = [
 ACTIVATIONS = [
     "relu",
     "tanh",
+    "prelu",
 ]
 
 MAX_IC50_VALUES = [
@@ -48,30 +50,45 @@ MAX_IC50_VALUES = [
     20000,
 ]
 
+EMBEDDING_SIZES = [
+    0,
+    64,
+]
+
 
 def generate_all_model_configs(
-        embedding_sizes=[0, 64],
-        n_training_epochs=100,
-        max_dropout=0.25):
+        hidden_layer_sizes=HIDDEN_LAYER_SIZES,
+        embedding_sizes=EMBEDDING_SIZES,
+        init_methods=INITILIZATION_METHODS,
+        activations=ACTIVATIONS,
+        n_training_epochs_values=[100],
+        n_pretrain_epochs_values=[0, 10],
+        minibatch_sizes=[128],
+        dropout_values=[0.0, 0.25],
+        max_ic50_values=[5000, 20000],
+        losses=["mse"]):
     configurations = []
-    for activation in ACTIVATIONS:
-        for loss in ["mse"]:
-            for init in INITILIZATION_METHODS:
-                for n_pretrain_epochs in [0, 10]:
-                    for hidden_layer_size in HIDDEN1_LAYER_SIZES:
-                        for embedding_size in embedding_sizes:
-                            for dropout in [0, max_dropout]:
-                                for max_ic50 in MAX_IC50_VALUES:
-                                    config = ModelConfig(
-                                        embedding_size=embedding_size,
-                                        hidden_layer_size=hidden_layer_size,
-                                        activation=activation,
-                                        init=init,
-                                        loss=loss,
-                                        dropout_probability=dropout,
-                                        n_pretrain_epochs=n_pretrain_epochs,
-                                        n_epochs=n_training_epochs,
-                                        max_ic50=max_ic50)
-                                    print(config)
-                                    configurations.append(config)
+    for activation in activations:
+        for loss in losses:
+            for init in init_methods:
+                for n_pretrain in n_pretrain_epochs_values:
+                    for n_training_epochs in n_training_epochs_values:
+                        for hidden in hidden_layer_sizes:
+                            for embedding_size in embedding_sizes:
+                                for dropout in dropout_values:
+                                    for max_ic50 in max_ic50_values:
+                                        for minibatch_size in minibatch_sizes:
+                                            config = ModelConfig(
+                                                embedding_size=embedding_size,
+                                                hidden_layer_size=hidden,
+                                                activation=activation,
+                                                init=init,
+                                                loss=loss,
+                                                dropout_probability=dropout,
+                                                n_pretrain_epochs=n_pretrain,
+                                                n_epochs=n_training_epochs,
+                                                max_ic50=max_ic50,
+                                                minibatch_size=minibatch_size)
+                                            print(config)
+                                            configurations.append(config)
     return configurations
